backend/pyproject.toml
[project]
name = "cco-assistant-backend"
version = "0.1.0"
description = "CCO Backend - Decision Outsourcing Learning Assistant"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "openai>=1.55.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  # v0.4: 多用户认证
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
